Working google auth for both frameworks
