+++ mcp-config.json
@@ -1,0 +1,13 @@
+{
+  "mcpServers": {
+    "actingweb-demo": {
+      "command": "python",
+      "args": ["fastapi_application.py"],
+      "env": {
+        "APP_HOST_FQDN": "localhost:5000",
+        "APP_HOST_PROTOCOL": "http://",
+        "LOG_LEVEL": "INFO"
+      }
+    }
+  }
+}
