+++ pyproject.toml
@@ -16,8 +16,8 @@ python = "^3.11"
 jinja2 = ">=3.1.2"
 flask = ">=2.3.0,<4.0.0"
 uwsgi = ">=2.0.23"
-actingweb = ">=3.1"
-#actingweb = { path = "../actingweb", develop = true }
+#actingweb = ">=3.1"
+actingweb = { path = "../actingweb", develop = true }
 
 [tool.poetry.group.dev.dependencies]
 pydevd-pycharm = "*"
