+++ templates/aw-root-factory.html
@@ -1,14 +1,52 @@
 <html>
+  <head>
+    <title>ActingWeb Demo - Create Actor</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <style>
+      body { font-family: Arial, sans-serif; margin: 40px; background-color: #f5f5f5; }
+      .container { max-width: 500px; margin: 0 auto; background: white; padding: 30px; border-radius: 8px; box-shadow: 0 2px 10px rgba(0,0,0,0.1); }
+      h1 { color: #333; margin-bottom: 20px; }
+      .description { margin-bottom: 25px; color: #666; line-height: 1.5; }
+      .form-group { margin-bottom: 20px; }
+      label { display: block; margin-bottom: 5px; font-weight: bold; color: #333; }
+      input[type="email"] { width: 100%; padding: 10px; border: 1px solid #ddd; border-radius: 4px; font-size: 16px; }
+      input[type="submit"] { background-color: #4285f4; color: white; padding: 12px 20px; border: none; border-radius: 4px; cursor: pointer; font-size: 16px; }
+      input[type="submit"]:hover { background-color: #3367d6; }
+      .error { color: #d93025; margin-bottom: 15px; padding: 10px; background-color: #fce8e6; border-radius: 4px; }
+      .oauth-info { background-color: #e8f0fe; padding: 15px; border-radius: 4px; margin-bottom: 20px; font-size: 14px; }
+    </style>
+  </head>
   <body>
-  <h1>Welcome to the ActingWeb Demo</h1>
-  To register an actor, fill out the form below.<br>
-    <form action="/" method="post">
-      <div><label>
-          Your  email:
-          <input type="text" name="creator" maxlength="100"/>
-      </label></div>
-      <div><input type="submit" value="Create new actor"></div>
-    </form>
-
+    <div class="container">
+      <h1>Welcome to ActingWeb Demo</h1>
+      
+      <div class="oauth-info">
+        📧 Enter your email address to login to or create your ActingWeb actor. You'll be redirected to authenticate with your OAuth2 provider (Google/GitHub) to verify your identity.
+      </div>
+      
+      {% if error %}
+        <div class="error">{{ error }}</div>
+      {% endif %}
+      
+      <div class="description">
+        ActingWeb actors are distributed micro-services that can securely share data and communicate with each other. Each actor has its own unique URL and properties.
+      </div>
+      
+      <form action="/" method="post">
+        <div class="form-group">
+          <label for="creator">Your Email Address:</label>
+          <input type="email" name="creator" id="creator" maxlength="100" required 
+                 placeholder="Enter your email address" />
+        </div>
+        <div class="form-group">
+          <input type="submit" value="Login or create new actor" />
+        </div>
+      </form>
+      
+      <div style="margin-top: 30px; font-size: 12px; color: #999; text-align: center;">
+        Powered by ActingWeb | <a href="/docs" style="color: #4285f4;">API Documentation</a>
+      </div>
+    </div>
   </body>
 </html>
